pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentpay-x402"
version = "0.1.1"
description = "Pay-per-call crypto data for AI agents — x402 + USDC on Stellar or Base"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "romudille", email = "velvetvau@gmail.com" }]
requires-python = ">=3.10"
keywords = ["x402", "stellar", "ai-agents", "crypto", "micropayments", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "stellar-sdk>=10.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage    = "https://github.com/romudille-bit/agentpay"
Repository  = "https://github.com/romudille-bit/agentpay"
Issues      = "https://github.com/romudille-bit/agentpay/issues"

[tool.setuptools.packages.find]
include = ["agentpay*"]
